+++ src/main/java/woowacourse/shoppingcart/dto/request/DeleteCustomerDto.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.request;
 
 import javax.validation.constraints.NotBlank;
