+++ src/main/java/woowacourse/shoppingcart/service/CustomerService.java
@@ -5,10 +5,10 @@ import org.springframework.stereotype.Service;
 import woowacourse.auth.support.PasswordEncoder;
 import woowacourse.shoppingcart.dao.CustomerDao;
 import woowacourse.shoppingcart.domain.Customer.Customer;
-import woowacourse.shoppingcart.dto.CustomerDto;
-import woowacourse.shoppingcart.dto.DeleteCustomerDto;
-import woowacourse.shoppingcart.dto.SignUpDto;
-import woowacourse.shoppingcart.dto.UpdateCustomerDto;
+import woowacourse.shoppingcart.dto.response.CustomerDto;
+import woowacourse.shoppingcart.dto.request.DeleteCustomerDto;
+import woowacourse.shoppingcart.dto.request.SignUpDto;
+import woowacourse.shoppingcart.dto.request.UpdateCustomerDto;
 import woowacourse.shoppingcart.exception.DuplicateNameException;
 import woowacourse.shoppingcart.exception.InvalidCustomerException;
 import woowacourse.shoppingcart.exception.PasswordMismatchException;
