+++ src/test/java/woowacourse/shoppingcart/service/CustomerServiceTest.java
@@ -8,10 +8,10 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.annotation.DirtiesContext;
 import woowacourse.auth.support.PasswordEncoder;
-import woowacourse.shoppingcart.dto.CustomerDto;
-import woowacourse.shoppingcart.dto.DeleteCustomerDto;
-import woowacourse.shoppingcart.dto.SignUpDto;
-import woowacourse.shoppingcart.dto.UpdateCustomerDto;
+import woowacourse.shoppingcart.dto.response.CustomerDto;
+import woowacourse.shoppingcart.dto.request.DeleteCustomerDto;
+import woowacourse.shoppingcart.dto.request.SignUpDto;
+import woowacourse.shoppingcart.dto.request.UpdateCustomerDto;
 import woowacourse.shoppingcart.exception.DuplicateNameException;
 import woowacourse.shoppingcart.exception.InvalidCustomerException;
